refactor(tui): modularize runtime, app state, input, render, and log handling
