+++ crates/sw_galaxy_map_cli/src/cli/shell.rs
@@ -1,7 +1,7 @@
-use crate::cli::tui;
+use crate::tui::run_tui;
 
 pub(crate) fn run_interactive_shell(db_arg: Option<String>) -> anyhow::Result<()> {
-    tui::run_tui(db_arg).map_err(Into::into)
+    run_tui(db_arg).map_err(Into::into)
 }
 
 pub(crate) fn split_args(line: &str) -> anyhow::Result<Vec<String>> {
